Move some more tests
