+++ crates/polars-io/src/csv/read/splitfields.rs
@@ -329,3 +329,26 @@ mod inner {
 }
 
 pub(crate) use inner::SplitFields;
+
+#[cfg(test)]
+mod test {
+    use super::SplitFields;
+
+    #[test]
+    fn test_splitfields() {
+        let input = "\"foo\",\"bar\"";
+        let mut fields = SplitFields::new(input.as_bytes(), b',', Some(b'"'), b'\n');
+
+        assert_eq!(fields.next(), Some(("\"foo\"".as_bytes(), true)));
+        assert_eq!(fields.next(), Some(("\"bar\"".as_bytes(), true)));
+        assert_eq!(fields.next(), None);
+
+        let input2 = "\"foo\n bar\";\"baz\";12345";
+        let mut fields2 = SplitFields::new(input2.as_bytes(), b';', Some(b'"'), b'\n');
+
+        assert_eq!(fields2.next(), Some(("\"foo\n bar\"".as_bytes(), true)));
+        assert_eq!(fields2.next(), Some(("\"baz\"".as_bytes(), true)));
+        assert_eq!(fields2.next(), Some(("12345".as_bytes(), false)));
+        assert_eq!(fields2.next(), None);
+    }
+}
